requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
SQLAlchemy==1.4.48
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
pandas==2.1.3 # Ya lo tenías, pero lo confirmo
openpyxl==3.1.2 # Ya lo tenías, pero lo confirmo